[DEL] dependency on local files
